st/prober: always try to resolve interface ID
The commit 445829f6ce84 ("st/prober: Add check for virtual hidraw
devices") added a check for ID_PATH. However, this is a udev property
that is set by rules and may not always be set for actual physical
devices.

Quite likely, one could just move the check into the HIDRAW_BUS_USB.
But, it is also perfectly save to ignore the error if the device does
not have a parent in the USB subsystem. So do that for now.

Fixes: 445829f6ce84 ("st/prober: Add check for virtual hidraw device")